Moved dialog/reward utils.
