+++ src/engine/core/utils/object/index.ts
@@ -11,7 +11,6 @@ export * from "@/engine/core/utils/object/object_setup";
 export * from "@/engine/core/utils/object/object_smart_cover";
 export * from "@/engine/core/utils/object/object_sound";
 export * from "@/engine/core/utils/object/object_state";
-export * from "@/engine/core/utils/object/object_task_reward";
 export * from "@/engine/core/utils/object/object_weapon";
 export * from "@/engine/core/utils/object/object_wounds";
 export * from "@/engine/core/utils/object/object_zone";
