+++ src/engine/core/utils/reward.ts
@@ -9,7 +9,6 @@ import {
 import { abort, assert, assertDefined } from "@/engine/core/utils/assertion";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { spawnItemsForObject } from "@/engine/core/utils/spawn";
-import { ACTOR_ID } from "@/engine/lib/constants/ids";
 import { ammo, TAmmoItem } from "@/engine/lib/constants/items/ammo";
 import { medkits, TMedkit } from "@/engine/lib/constants/items/drugs";
 import { ClientObject, LuaArray, Optional, TCount, TName, TNumberId, TSection } from "@/engine/lib/types";
@@ -49,17 +48,6 @@ export function transferMoneyFromActor(to: ClientObject, amount: TCount): void {
     direction: ENotificationDirection.OUT,
     amount,
   });
-}
-
-/**
- * From two possible speakers pick NPC one, omit actor.
- *
- * @param first - possible non-actor speaker
- * @param second - possible non-actor speaker
- * @returns non-actor game object picked from parameters
- */
-export function getNpcSpeaker(first: ClientObject, second: ClientObject): ClientObject {
-  return second.id() === ACTOR_ID ? first : second;
 }
 
 /**
