+++ src/engine/scripts/declarations/dialogs/dialogs.ts
@@ -10,21 +10,19 @@ import { ISchemeMeetState } from "@/engine/core/schemes/stalker/meet";
 import { updateObjectMeetAvailability } from "@/engine/core/schemes/stalker/meet/utils";
 import { ISchemeWoundedState } from "@/engine/core/schemes/stalker/wounded";
 import { extern } from "@/engine/core/utils/binding";
+import { getNpcSpeaker } from "@/engine/core/utils/dialog";
 import { createGameAutoSave } from "@/engine/core/utils/game";
 import { giveInfo, hasAlifeInfo } from "@/engine/core/utils/info_portion";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import {
-  actorHasMedKit,
   enableObjectWoundedHealing,
-  getActorAvailableMedKit,
-  getNpcSpeaker,
   getObjectCommunity,
   isObjectInjured,
   isObjectInSmartTerrain,
   isObjectWounded,
   isStalkerAlive,
-  transferItemsFromActor,
 } from "@/engine/core/utils/object";
+import { actorHasMedKit, getActorAvailableMedKit, transferItemsFromActor } from "@/engine/core/utils/reward";
 import { communities } from "@/engine/lib/constants/communities";
 import { infoPortions } from "@/engine/lib/constants/info_portions";
 import { drugs, TMedkit } from "@/engine/lib/constants/items/drugs";
