+++ src/engine/scripts/declarations/dialogs/dialogs_jupiter.ts
@@ -5,19 +5,19 @@ import { getPortableStoreValue } from "@/engine/core/database/portable_store";
 import { TreasureManager } from "@/engine/core/managers/treasures";
 import { AnomalyZoneBinder } from "@/engine/core/objects/binders/zones/AnomalyZoneBinder";
 import { extern, getExtern } from "@/engine/core/utils/binding";
+import { getNpcSpeaker } from "@/engine/core/utils/dialog";
 import { disableInfo, giveInfo, hasAlifeInfo } from "@/engine/core/utils/info_portion";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { isObjectInSmartTerrain } from "@/engine/core/utils/object";
+import { getObjectsRelationSafe, isActorEnemyWithFaction } from "@/engine/core/utils/relation";
 import {
-  getNpcSpeaker,
   giveItemsToActor,
   giveMoneyToActor,
   isObjectName,
   transferItemsFromActor,
   transferItemsToActor,
   transferMoneyFromActor,
-} from "@/engine/core/utils/object/object_task_reward";
-import { getObjectsRelationSafe, isActorEnemyWithFaction } from "@/engine/core/utils/relation";
+} from "@/engine/core/utils/reward";
 import { communities } from "@/engine/lib/constants/communities";
 import { ACTOR_ID } from "@/engine/lib/constants/ids";
 import { infoPortions, TInfoPortion } from "@/engine/lib/constants/info_portions/info_portions";
