+++ src/engine/scripts/declarations/dialogs/dialogs_pripyat.ts
@@ -1,14 +1,14 @@
 import { registry } from "@/engine/core/database";
 import { extern, getExtern } from "@/engine/core/utils/binding";
+import { getNpcSpeaker } from "@/engine/core/utils/dialog";
 import { disableInfo, giveInfo, hasAlifeInfo } from "@/engine/core/utils/info_portion";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import {
-  getNpcSpeaker,
   giveMoneyToActor,
   transferItemsFromActor,
   transferItemsToActor,
   transferMoneyFromActor,
-} from "@/engine/core/utils/object/object_task_reward";
+} from "@/engine/core/utils/reward";
 import { infoPortions, TInfoPortion } from "@/engine/lib/constants/info_portions/info_portions";
 import { TInventoryItem } from "@/engine/lib/constants/items";
 import { ammo, TAmmoItem } from "@/engine/lib/constants/items/ammo";
